refactor: 삭제 유효성 검증 추가

+++ src/main/java/roomescape/service/ReservationService.java
@@ -109,6 +109,9 @@ public class ReservationService {
     }
 
     public void removeReservation(long id) {
+        if (reservationRepository.existsById(id)) {
+            throw new NotFoundException("reservation");
+        }
         reservationRepository.deleteById(id);
     }
 }
